# java/recruitment-api/src/main/java/com/talent/recruitment/web/dto/PageInfo.java
package com.talent.recruitment.web.dto;

import lombok.AllArgsConstructor;
import lombok.Builder;
import lombok.Data;
import lombok.NoArgsConstructor;

@Data
@Builder
@NoArgsConstructor
@AllArgsConstructor
public class PageInfo {
    private int number;
    private int size;
    private long totalElements;
    private int totalPages;
}